requirements.txt
anyio
beautifulsoup4
bs4
diffusers
fastapi
ffmpeg-python
future
GitPython
gradio
httpx
huggingface-hub
icetk
lxml
Markdown
markdown-it-py
matplotlib
nltk
numpy
opencv-python
pandas
pdfplumber
Pillow
protobuf==3.20.0
python-docx
PyYAML
requests
rich
selenium
sse-starlette
sseclient
streamlit
streamlit-chat
supervision
torch
torchaudio
torchvision
tqdm
transformers
urllib3
uvicorn
websockets